fix(architecture): relocate SoftMotion axis-naming to middleware/shared/utils
PR #928 introduced imports of generate-softmotion.ts's pure axis-naming
helpers (sanitizeAxisName, softMotionAxisNames, serializeSoftMotionAxisGlobalsToST,
isValidIecIdentifier) and cia402.ts directly from three frontend/services/
and frontend/store/ files, plus two frontend/components/ call sites (missed
by the checker's @root/ alias blind spot) — all violating the
services/store/components -> backend-shared layer rule and failing
validate:arch on development for every PR since.

Rather than special-casing the violation with a KNOWN_EXCEPTIONS entry,
relocated the pure logic to middleware/shared/utils/ethercat/ — the
existing home for domain logic reachable from every layer including the
compile pipeline (same relationship target-capabilities/ and library/
already have to backend/shared/compile/pipeline.ts). backend/shared/ethercat/
generate-softmotion.ts now only owns the actual codegen
(generateSoftMotionArtifacts/injectAxisExternals), importing collectAxes
from the new utils module.

Also fixed a pre-existing prettier violation from the same PR
(preprocess-pous.test.ts) that was failing the format check on development.
(openplc-web's matching softmotion-e2e.test.ts fix does not apply here —
that test is web-adapter-specific and doesn't exist in this repo.)
