+++ src/backend/shared/utils/PLC/__tests__/preprocess-pous.test.ts
@@ -331,8 +331,28 @@ describe('preprocessPous — mixed', () => {
                 config: {},
                 cia402: { enabled: true, scaleNum: 1, scaleDenom: 1, scaleFactor: 1 },
                 channelInfo: [
-                  { channelId: 'c1', name: 'Controlword', direction: 'output', pdoIndex: '0x1600', entryIndex: '0x6040', entrySubIndex: '0x0', dataType: 'UINT', bitLen: 16, iecType: 'UINT' },
-                  { channelId: 'c2', name: 'Statusword', direction: 'input', pdoIndex: '0x1A00', entryIndex: '0x6041', entrySubIndex: '0x0', dataType: 'UINT', bitLen: 16, iecType: 'UINT' },
+                  {
+                    channelId: 'c1',
+                    name: 'Controlword',
+                    direction: 'output',
+                    pdoIndex: '0x1600',
+                    entryIndex: '0x6040',
+                    entrySubIndex: '0x0',
+                    dataType: 'UINT',
+                    bitLen: 16,
+                    iecType: 'UINT',
+                  },
+                  {
+                    channelId: 'c2',
+                    name: 'Statusword',
+                    direction: 'input',
+                    pdoIndex: '0x1A00',
+                    entryIndex: '0x6041',
+                    entrySubIndex: '0x0',
+                    dataType: 'UINT',
+                    bitLen: 16,
+                    iecType: 'UINT',
+                  },
                 ],
                 channelMappings: [
                   { channelId: 'c1', iecLocation: '%QW0' },
